Moved cstddef tests
