+++ kernel/UnitTests.cpp
@@ -21,24 +21,6 @@ namespace
     using namespace ::UnitTests;
 
     ///////////////////////////////////////////////////////////////////////////
-    // cstddef/stddef.h tests
-    ///////////////////////////////////////////////////////////////////////////
-    struct TestOffsetStruct
-    {
-        uint8_t ExpectedAt0;
-        uint32_t ExpectedAt4;
-        uint64_t ExpectedAt8;
-    };
-
-    static_assert(sizeof(std::size_t) == 8, "Unexpected size_t size");
-    static_assert(sizeof(std::ptrdiff_t) == 8, "Unexpected ptrdiff_t size");
-    static_assert(NULL == 0, "Unexpected NULL value");
-    static_assert(alignof(std::max_align_t) == 16, "Unexpected max_align_t alignment");
-    static_assert(offsetof(TestOffsetStruct, ExpectedAt0) == 0, "Unexpected offsetof first member");
-    static_assert(offsetof(TestOffsetStruct, ExpectedAt4) == 4, "Unexpected offsetof second member");
-    static_assert(offsetof(TestOffsetStruct, ExpectedAt8) == 8, "Unexpected offsetof third member");
-
-    ///////////////////////////////////////////////////////////////////////////
     // cstdint/stdint.h tests
     ///////////////////////////////////////////////////////////////////////////
     static_assert(sizeof(std::int8_t) == 1, "Unexpected int8_t size");
@@ -607,6 +589,7 @@ namespace UnitTests
 
         KernelStdlib::Bitset::Run();
         // No runtime tests for climits
+        // No runtime tests for cstddef
         
         StdMoveTest();
         StdForwardTest();
