+++ kernel/UnitTests/KernelStdlib/CMakeLists.txt
@@ -2,4 +2,5 @@ target_sources(kernel8.elf
     PRIVATE
         BitsetTests.h BitsetTests.cpp
         CLimitsTests.cpp
+        CStdDefTests.cpp
 )
